feat: add abbybot news blueprint and implement news routes
