+++ abbybot_website/app/__init__.py
@@ -1,5 +1,3 @@
-# app/__init__.py
-
 from flask import Flask
 from flask_mail import Mail
 from dotenv import load_dotenv
@@ -39,7 +37,7 @@ def create_app():
     from app.routes.bot_policies import bot_policies_bp
     from app.routes.user_responsibilities import user_responsibilities_bp
     from app.routes.commands import commands_bp
-
+    from app.routes.abbybot_news import abbybot_news_bp
     from app.routes.handlers.error_handlers import error_handlers_bp
 
     app.register_blueprint(main_bp)
@@ -54,5 +52,6 @@ def create_app():
     app.register_blueprint(bot_policies_bp)
     app.register_blueprint(user_responsibilities_bp)
     app.register_blueprint(commands_bp)
+    app.register_blueprint(abbybot_news_bp)
     
     return app
